pyproject.toml
[project]
name = "padelbot"
version = "1.5.0"
description = "An app to simplify the life of managing the Equinor BIL Trondheim padel group"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "configparser>=7.2.0",
    "naco-backend-client",
    "python-dotenv>=1.2.2",
    "spond>=1.2.0",
    "starlette>=1.0.0",
    "types-aiofiles>=25.1.0.20260409",
    "uvicorn>=0.46.0",
]

[tool.uv.sources]
naco-backend-client = { path = "naco-backend-client" }

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "openapi-python-client>=0.28.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
    "types-python-dateutil>=2.9.0.20260408",
]

[tool.pytest.ini_options]
pythonpath = "."
